+++ tests/doc/capi/chfl_frame/set_index.c
@@ -2,18 +2,17 @@
 // Copyright (C) Guillaume Fraux and contributors -- BSD license
 
 #include <chemfiles.h>
-#include <stdlib.h>
 #include <assert.h>
 
 int main(void) {
     // [example]
     CHFL_FRAME* frame = chfl_frame();
 
-    chfl_frame_set_step(frame, 678);
+    chfl_frame_set_index(frame, 678);
 
-    uint64_t step = 0;
-    chfl_frame_step(frame, &step);
-    assert(step == 678);
+    uint64_t index = 0;
+    chfl_frame_index(frame, &index);
+    assert(index == 678);
 
     chfl_free(frame);
     // [example]